Add authorization checks to updatePost and deletePost

+++ controllers/feed.js
@@ -133,6 +133,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404; //* Not found!
         throw error; //todo When throw inside a then, the next catch block is reached
       }
+      //! Check if creatorID is equal to the Id of currently loggedin user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -164,6 +170,11 @@ exports.deletePost = (req, res, next) => {
         const error = new Error("Could not find post.");
         error.statusCode = 404; //* Not found!
         throw error; //todo When throw inside a then, the next catch block is reached
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
       }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
